fix: fixing references and moving convert
